Change goToPrev to handle multiple function steps
Current implementation steps back one step, rolls it back, then
continues at the step before that.

Problem arises when there are two or more function steps in a row. When
it tries to back up over them, it rolls back the highest index function,
sets current to the previous function, then continues on, executing the
two functions, then plunking you back in the title/element step you were
on.

This change iterates backwards and rolls back until we hit a
title/element to handle this sequence.